+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## ????
 - Provide default HTTP options
 - Default to HTTP/1.1
+- Run PHP lint as part of the tests
 
 ## 0.0.2 - 2023-04-08
 - Rename to HTTP