+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Default to HTTP/1.1
 - Run PHP lint as part of the tests
 - Use the `ignore_errors` option instead of the error handler to capture response error conditions
+- New test: `get: fail, 500`
 
 ## 0.0.2 - 2023-04-08
 - Rename to HTTP