+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - New test: `get: timeout`
 - New public method: request().  This centralizes the file_get_contents() call into one place.
 - Update the get() and post() methods to use the new request() method
+- Track the total time a request took
 
 ## 0.0.2 - 2023-04-08
 - Rename to HTTP