+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - New public method: request().  This centralizes the file_get_contents() call into one place.
 - Update the get() and post() methods to use the new request() method
 - Track the total time a request took
+- New test: `get: slow`, make sure total time tracking works
 
 ## 0.0.2 - 2023-04-08
 - Rename to HTTP