+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## ????
+## 0.0.3 - 2023-04-10
 - Provide default HTTP options
 - Default to HTTP/1.1
 - Run PHP lint as part of the tests